better delete

+++ app/controllers/modals/delete-list.js
@@ -6,17 +6,22 @@ export default Ember.Controller.extend({
             return this.send('closeModal');
         },
         deleteList: function(){
+            var loggedInId = this.get('session.user_id');
+            var ownerId = this.get('model.user.id');
 
-            var that = this;
-            var list = this.get('model');
-            list.destroyRecord().then(function(){
-                // Success callback
-                that.send('closeModal');
-                //that.transitionToRoute('lists');
-            }, function(error) {
-                console.log('tried save, ERROR', error);
-                return false;
-            });
+            if (loggedInId && (loggedInId === ownerId)){
+                var that = this;
+                var list = this.get('model');
+                var user = this.get('model.user');
+                list.destroyRecord().then(function(){
+                    // Success callback
+                    that.send('closeModal');
+                    that.transitionToRoute('user',user);
+                }, function(error) {
+                    console.log('tried save, ERROR', error);
+                    return false;
+                });
+            }
             return;  
         }
     }
